refactor(dashboard): 로딩 스켈레톤을 별도 컴포넌트로 분리
DashboardView.tsx 한 파일에 컴포넌트가 둘 있어서 "파일명은 컴포넌트명과
일치" 규칙과 어긋났습니다. 참가자 화면이 이미 components/live/LiveSkeleton.tsx로
떼어 놓은 선례를 따랐습니다.

내용은 그대로 옮겼고 렌더 결과는 같습니다.

Refs #115
